requirements.txt
aiobotocore==2.13.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
attrs==23.2.0
bokeh==3.5.0
botocore==1.34.131
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cramjam==2.8.3
dask==2024.7.1
dask-expr==1.1.9
decorator==5.1.1
distributed==2024.7.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gcsfs==2024.6.1
google-api-core==2.19.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
idna==3.7
importlib_metadata==8.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
locket==1.0.0
lz4==4.3.3
MarkupSafe==2.1.5
modin==0.31.0
msgpack==1.0.8
multidict==6.0.5
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
partd==1.4.2
pillow==10.4.0
polars==1.2.1
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
py4j==0.10.9.7
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.33.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
rsa==4.9
six==1.16.0
sortedcontainers==2.4.0
tblib==3.0.0
toolz==0.12.1
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
xyzservices==2024.6.0
yarl==1.9.4
zict==3.0.0
zipp==3.19.2
